Refactoring: Introduce `CLICommand` to reuse CLI related logic
