+++ lib/hanami/assets/precompiler.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "open3"
-require "shellwords"
+require_relative "cli_command"
 
 module Hanami
   class Assets
@@ -10,27 +10,7 @@ module Hanami
     #
     # @since 0.1.0
     # @api private
-    class Precompiler
-      # @since 2.1.0
-      # @api private
-      attr_reader :config
-      private :config
-
-      # @since 2.1.0
-      # @api private
-      def initialize(config:)
-        @config = config
-
-        freeze
-      end
-
-      # @since 2.1.0
-      # @api private
-      def call
-        cmd, *args = cmd_with_args
-        execute(cmd, env, *args)
-      end
-
+    class Precompiler < CLICommand
       private
 
       # @since 2.1.0
@@ -46,11 +26,7 @@ module Hanami
       # @since 2.1.0
       # @api private
       def cmd_with_args
-        result = [
-          config.package_manager_executable,
-          config.package_manager_command,
-          config.executable
-        ]
+        result = super
 
         if config.subresource_integrity.any?
           result << "--"
@@ -58,35 +34,6 @@ module Hanami
         end
 
         result
-      end
-
-      # @since 2.1.0
-      # @api private
-      def env
-        ENV.to_h.merge(
-          "ESBUILD_ENTRY_POINTS" => entry_points,
-          "ESBUILD_OUTDIR" => destination
-        )
-      end
-
-      # @since 2.1.0
-      # @api private
-      def entry_points
-        config.entry_points.map do |entry_point|
-          escape(entry_point)
-        end.join(" ")
-      end
-
-      # @since 2.1.0
-      # @api private
-      def destination
-        escape(config.destination)
-      end
-
-      # @since 2.1.0
-      # @api private
-      def escape(str)
-        Shellwords.shellescape(str)
       end
     end
   end
