+++ lib/hanami/assets/watcher.rb
@@ -1,33 +1,12 @@
 # frozen_string_literal: true
 
-require "shellwords"
+require_relative "cli_command"
 
 module Hanami
   class Assets
     # @since 2.1.0
     # @api private
-    class Watcher
-      # @since 2.1.0
-      # @api private
-      attr_reader :config
-      private :config
-
-      # @since 2.1.0
-      # @api private
-      def initialize(config:)
-        super()
-        @config = config
-
-        freeze
-      end
-
-      # @since 2.1.0
-      # @api private
-      def call
-        cmd, *args = cmd_with_args
-        execute(cmd, env, *args)
-      end
-
+    class Watcher < CLICommand
       private
 
       # @since 2.1.0
@@ -45,42 +24,11 @@ module Hanami
       # @since 2.1.0
       # @api private
       def cmd_with_args
-        [
-          config.package_manager_executable,
-          config.package_manager_command,
-          config.executable,
-          "--",
-          "--watch"
-        ]
-      end
-
-      # @since 2.1.0
-      # @api private
-      def env
-        ENV.to_h.merge({
-                         "ESBUILD_ENTRY_POINTS" => entry_points,
-                         "ESBUILD_OUTDIR" => destination
-                       })
-      end
-
-      # @since 2.1.0
-      # @api private
-      def entry_points
-        config.entry_points.map do |entry_point|
-          escape(entry_point)
-        end.join(" ")
-      end
-
-      # @since 2.1.0
-      # @api private
-      def destination
-        escape(config.destination)
-      end
-
-      # @since 2.1.0
-      # @api private
-      def escape(str)
-        Shellwords.shellescape(str)
+        super +
+          [
+            "--",
+            "--watch"
+          ]
       end
     end
   end
